+++ res/raw/phong_material_vertex.glsl
@@ -4,11 +4,16 @@ uniform mat4 uMVPMatrix;
 uniform mat3 uNMatrix;
 uniform mat4 uMMatrix;
 uniform mat4 uVMatrix;
+#ifdef USE_SINGLE_COLOR
+uniform vec4 uSingleColor;
+#endif
 
 attribute vec4 aPosition;
 attribute vec3 aNormal;
 attribute vec2 aTextureCoord;
+#ifdef USE_VERTEX_COLOR
 attribute vec4 aColor;
+#endif
 
 varying vec2 vTextureCoord;
 varying vec3 vNormal;
@@ -55,6 +60,11 @@ M_SKELETAL_ANIM_VERTEX_MATRIX
 
 %LIGHT_CODE%
 
-   vColor = aColor;
+//vColor = aColor;
+
+#ifdef USE_SINGLE_COLOR
+	vColor = uSingleColor;
+#endif
+   
 M_FOG_VERTEX_DENSITY
 }