+++ src/rajawali/bounds/BoundingBox.java
@@ -69,7 +69,7 @@ public class BoundingBox implements IBoundingVolume {
 		if(mVisualBox == null) {
 			mVisualBox = new Cube(1);
 			mVisualBox.setMaterial(new SimpleMaterial());
-			mVisualBox.getMaterial().setUseColor(true);
+			mVisualBox.getMaterial().setUseSingleColor(true);
 			mVisualBox.setColor(mBoundingColor.get());
 			mVisualBox.setDrawingMode(GLES20.GL_LINE_LOOP);
 			mVisualBox.setDoubleSided(true);